refactor: auto-split test_external_integration.py
